Actulizacion
Ya mira que no se pase del maximo de celdas de carros y motos

+++ src/main/java/dominio/ConstantesParqueadero.java
@@ -1,8 +1,8 @@
 package dominio;
 
 public class ConstantesParqueadero {
-	public static final int MAXIMO_CARROS =20;
-	public static final int MAXIMO_MOTOS =10;
+	public static final long MAXIMO_CARROS =20;
+	public static final long MAXIMO_MOTOS =10;
 	public static final double VALOR_HORA_CARRO =1000d;
 	public static final double VALOR_DIA_CARRO =8000d;
 	public static final double VALOR_DIA_MOTO =600d;
